Fix duplicate review issue

+++ final_project/api/routers/review.py
@@ -1,22 +1,16 @@
 from typing import Annotated
+from beanie import PydanticObjectId
 from fastapi import APIRouter, Path, HTTPException, status
 
 from models.review import Review, ReviewRequest
 
 review_router = APIRouter()
 
-max_id: int = (
-    0  # NOT IDEAL because if you stop session and reload, then you open the door for duplicate ID's. Try doing a query of the database instead to get id numbers.
-)
-
 
 @review_router.post("", status_code=status.HTTP_201_CREATED)
 async def add_review(review: ReviewRequest) -> Review:
-    global max_id
-    max_id += 1  # auto increment ID
 
     newReview = Review(
-        id=max_id,
         restaurant=review.restaurant,
         rating=review.rating,
         description=review.description,
@@ -31,7 +25,7 @@ async def get_reviews() -> list[Review]:
 
 
 @review_router.get("/{id}")
-async def get_review_by_id(id: int = Path(..., title="default")) -> Review:
+async def get_review_by_id(id: PydanticObjectId) -> Review:
     review = await Review.get(id)
     if review:
         return review
@@ -42,7 +36,7 @@ async def get_review_by_id(id: int = Path(..., title="default")) -> Review:
 
 
 @review_router.put("/{id}")
-async def update_review(review: ReviewRequest, id: int) -> dict:
+async def update_review(review: ReviewRequest, id: PydanticObjectId) -> dict:
     existing_review = await Review.get(id)
     if existing_review:
         existing_review.restaurant = review.restaurant
@@ -55,7 +49,12 @@ async def update_review(review: ReviewRequest, id: int) -> dict:
 
 
 @review_router.delete("/{id}")
-async def delete_review(id: int) -> dict:
+async def delete_review(id: PydanticObjectId) -> dict:
     review = await Review.get(id)
-    await review.delete()
-    return {"message": f"The review with ID={id} has been deleted."}
+    if review:
+        await review.delete()
+        return {"message": f"The review with ID={id} has been deleted."}
+    raise HTTPException(
+        status_code=status.HTTP_404_NOT_FOUND,
+        detail=f"The movie with ID={id} is not found.",
+    )
